Rather than passing around a pointer, reference or shared_ptr of ios (rather indiscriminately), use a static method instead.
The original idea was that a monster would not have an ios object, so that actions applying to a monster would not be forwareded to the player.
But there are too many exceptions to this to be useful, and it just makes for messy code.
I think this way will be much simpler.

- RL