Armour can now be damaged or destroyed by combat.
Every time a monster successfully hits you, an outermost piece of armour will take 1 damage, unless it is proofed or inapplicable.

This is roughly as per the original plan, makes proofing and fixing worthwhile, and means metallic armour is no longer overpowered.

Note that most armour is fabric, which affords little protection and is impervious to bashing damage (which is the most common type, particularly with unarmed monsters), so metallic armour worn under a jerkin or robe is still very powerful.

- RL