https://github.com/kikigame/Tinn/issues/3
Debounce input keys.

If the browser sends an AJAX request while the previous request is being downloded/processed, then the server will process them in order as expected.
However, the browser may begin processing the second response before the first response is handled.
Because the first response hasn't finished yet, some of its data are still in the buffer. Thus the processing of the second request gets a traliing fragment of the first request, resulting in an "unknown instruction" error (data processed as instruction).

This commit means that any input key is ignored between the AJAX call being sent to the server and the response being read by the browser.
Page may crash if a non-200 response is received, but that could happen anyway.

- RL